requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
whitenoise==6.9.0
python-dotenv
drf-yasg
django_nose
coverage
